requirements.txt
paho-mqtt==1.5.0
pony==0.7.11
psycopg2-binary==2.8.4
